Reducing bundle size of the angular projects (web-frontend, events-frontend) using tree shaking in services, refactoring folder structure and setting up the environment
