+++ events-frontend/src/environments/environment.prod.ts
@@ -1,3 +1,4 @@
 export const environment = {
-  production: true
+  production: true,
+  baseURL: "http://localhost:5000"
 };
