+++ web-frontend/src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { WebsocketService } from './websocket.service';
+import { WebsocketService } from './services/websocket.service';
 
 @Component({
   selector: 'app-root',
