+++ web-frontend/src/app/views/message/message.component.scss
