+++ web-frontend/src/app/views/video-show/video-show.component.scss
